+++ src/modules/sns/SNSMessageDetailsDialog.cpp
@@ -1,8 +1,6 @@
 
-#include <QFileDialog>
 #include <modules/sns/SNSMessageDetailsDialog.h>
 #include "ui_SNSMessageDetailsDialog.h"
-#include "components/Toast.h"
 
 SNSMessageDetailsDialog::SNSMessageDetailsDialog(const QString &messageId, QWidget *parent) : QDialog(parent), _ui(new Ui::SNSMessageDetailsDialog), _messageId(messageId) {
     // Setup UI